requirements.txt
PyQt6>=6.4.0
pyqt-fluent-widgets>=1.2.0
mysql-connector-python>=8.0.28
requests>=2.28.1
aiohttp>=3.8.3
python-dotenv>=0.21.0
asyncio>=3.4.3 